Improve error messages for bad feature combos
The crate cannot function without at least one of the 'nightly' or 'std' features.

The 'unique-wrap-std' implicitly enables the 'std' feature,
but raises an error if the 'nightly' feature is not explicitly be enabled.
The 'nightly' requirement will be dropped if `#[feature(thread_id_value)]` is ever enabled.

Before this change, encountering this issue would trigger a nice `compile_error!`
with a descriptive error message, along with a bunch of other confusing looking errors.
After this change, only the single `compile_error!` is displayed.

This is implemented by adding stub implementations which call `unreachable!()`